rename parameters.py to stitcher_parameters.py
