+++ software/control/stitcher/stitcher_parameters.py
@@ -1,4 +1,4 @@
-# parameters.py
+# stitcher_parameters.py
 from dataclasses import dataclass
 from typing import Dict, Any
 import json
